Added beatmap delete dialog

+++ osu.Game/Screens/Select/PlaySongSelect.cs
@@ -50,6 +50,8 @@ namespace osu.Game.Screens.Select
 
         private List<BeatmapGroup> beatmapGroups;
 
+        private BeatmapDeleteDialog deleteDialog;
+
         private Footer footer;
 
         OsuScreen player;
@@ -122,7 +124,16 @@ namespace osu.Game.Screens.Select
                             PreferredPlayMode = playMode.Value
                         })).LoadAsync(Game, l => Push(player));
                     }
-                }
+                },
+                deleteDialog = new BeatmapDeleteDialog
+                {
+                    RelativeSizeAxes = Axes.Both,
+                    OnDelete = (b) =>
+                    {
+                        b.Dispose();
+                        database.Delete(b.BeatmapSetInfo);
+                    },
+                },
             };
 
             footer.AddButton(@"mods", colours.Yellow, null);
@@ -280,6 +291,7 @@ namespace osu.Game.Screens.Select
             //todo: change background in selectionChanged instead; support per-difficulty backgrounds.
             changeBackground(beatmap);
             carousel.SelectBeatmap(beatmap?.BeatmapInfo);
+            deleteDialog.Beatmap = beatmap;
         }
 
         /// <summary>
@@ -382,12 +394,20 @@ namespace osu.Game.Screens.Select
                     footer.StartButton.TriggerClick();
                     return true;
                 case Key.Delete:
-                    if (Beatmap != null)
+                    if (state.Keyboard.ShiftPressed)
                     {
-                        Beatmap.Dispose();
-                        database.Delete(Beatmap.BeatmapSetInfo);
+                        deleteDialog.Show();
+                        return true;
                     }
-                    return true;
+                    else
+                    {
+                        if (Beatmap != null)
+                        {
+                            Beatmap.Dispose();
+                            database.Delete(Beatmap.BeatmapSetInfo);
+                        }
+                        return true;
+                    }
             }
 
             return base.OnKeyDown(state, args);
